refactor/ create Model folder
