+++ src/classes/Model/BaseModel.ts
